fix: 오탈자 수정 및 도커 설정
