+++ compose.yml
@@ -1,0 +1,121 @@
+
+services:
+  # MySQL 서비스
+  my-mysql:
+    image: mysql:latest
+    container_name: my-mysql
+    environment:
+      MYSQL_ROOT_PASSWORD: password123
+    ports:
+      - "3306:3306"
+    volumes:
+      - /Users/jeongjuhyeon/개발/docker-mysql/mysql_data:/var/lib/mysql
+    networks:
+      - coupon-network
+    restart: always
+    healthcheck:
+      test: [ "CMD", "mysqladmin", "ping", "-h", "localhost" ]
+      interval: 30s
+      retries: 5
+      start_period: 30s
+      timeout: 10s
+
+  mysqld-exporter:
+    container_name: mysqld_exporter
+    image: prom/mysqld-exporter
+    environment:
+      - DATA_SOURCE_NAME="wngus:password123@(my-mysql:3306)/"
+    command:
+      - "--mysqld.username=wngus:password123"
+      - "--mysqld.address=my-mysql:3306"
+    ports:
+      - "9104:9104"
+    depends_on:
+      - my-mysql
+    restart: unless-stopped
+    networks:
+      - coupon-network  # 추가된 부분
+
+  # Redis (Cache Server) 서비스
+  cache-server:
+    image: redis
+    container_name: cache-server
+    ports:
+      - "6379:6379"
+    healthcheck:
+      test: [ "CMD", "redis-cli", "ping" ]
+      interval: 5s
+      retries: 10
+    networks:
+      - coupon-network
+
+  # Spring Boot 애플리케이션 서비스 (my-server)
+  my-server:
+    build: .
+    container_name: my-server
+    cpus: 2.0  # CPU 리소스 2개의 vCPU 할당
+    mem_limit: 2g  # 메모리 2GB로 설정
+    ports:
+      - "8080:8080"
+    depends_on:
+      - my-mysql
+      - cache-server
+    networks:
+      - coupon-network
+    privileged: true  # 호스트 자원 접근 허용
+    healthcheck:
+      test: [ "CMD", "curl", "-f", "http://localhost:8080/test" ]  # /test 엔드포인트로 건강 체크
+      interval: 30s
+      retries: 3
+      start_period: 10s
+      timeout: 10s
+
+  # Prometheus 서비스
+  prometheus:
+    image: prom/prometheus:latest
+    container_name: prometheus
+    volumes:
+      - ./prometheus.yml:/etc/prometheus/prometheus.yml
+    command:
+      - '--enable-feature=native-histograms'
+      - '--web.enable-remote-write-receiver'
+      - '--config.file=/etc/prometheus/prometheus.yml'
+    ports:
+      - "9090:9090"
+    networks:
+      - coupon-network
+
+  # Grafana 서비스
+  grafana:
+    image: grafana/grafana:latest
+    container_name: grafana
+    ports:
+      - "3000:3000"
+    networks:
+      - coupon-network
+
+  # K6 부하 테스트 서비스
+  k6:
+    image: grafana/k6
+    container_name: k6
+    volumes:
+      - ./k6-scripts:/scripts
+#    command: sleep infinity  # sleep 명령어로 컨테이너 계속 실행
+    command: run -o experimental-prometheus-rw /scripts/scripts.js
+    ports:
+      - "5665:5665"
+    environment:
+      - K6_WEB_DASHBOARD=true  # 대시보드 활성화
+      - K6_PROMETHEUS_RW_TREND_AS_NATIVE_HISTOGRAM=true
+      - K6_PROMETHEUS_RW_SERVER_URL=http://prometheus:9090/api/v1/write
+    extra_hosts:
+      - "host.docker.internal:host-gateway"
+    depends_on:
+      - prometheus
+#      - my-server  # my-server가 시작되기를 기다립니다.
+    networks:
+      - coupon-network
+
+networks:
+  coupon-network:
+    driver: bridge
