+++ src/main/java/com/personal_project/coupon/CouponApplication.java
@@ -1,13 +1,21 @@
 package com.personal_project.coupon;
 
+import jakarta.annotation.PostConstruct;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+import java.util.TimeZone;
 
 @SpringBootApplication
 @EnableJpaAuditing
 public class CouponApplication {
 
+    @PostConstruct
+    public void started() {
+        // timezone UTC 셋팅
+        TimeZone.setDefault(TimeZone.getTimeZone("Asia/Seoul"));
+    }
 	public static void main(String[] args) {
 		SpringApplication.run(CouponApplication.class, args);
 	}
