+++ src/main/java/com/personal_project/coupon/coupon/application/outputport/CouponIssueOutPort.java
@@ -1,6 +1,6 @@
 package com.personal_project.coupon.coupon.application.outputport;
 
-import com.personal_project.coupon.coupon.domain.entity.CouponIssue;
+import com.personal_project.coupon.coupon.domain.model.CouponIssue;
 
 public interface CouponIssueOutPort {
     boolean existByUserIdAndCouponId(Long memberId,Long couponId);
