+++ src/main/java/com/personal_project/coupon/coupon/application/outputport/CouponOutPort.java
@@ -1,6 +1,6 @@
 package com.personal_project.coupon.coupon.application.outputport;
 
-import com.personal_project.coupon.coupon.domain.entity.Coupon;
+import com.personal_project.coupon.coupon.domain.model.Coupon;
 
 import java.util.Optional;
