--- src/main/java/com/personal_project/coupon/coupon/application/usecase/CouponIssueFacade.java
@@ -1,5 +1,0 @@
-package com.personal_project.coupon.coupon.application.usecase;
-
-public interface CouponIssueFacade {
-    void issueCoupon(Long eventId, Long couponId, Long memberId);
-}
