--- src/main/java/com/personal_project/coupon/coupon/domain/enumeration/EventStatus.java
@@ -1,5 +1,0 @@
-package com.personal_project.coupon.coupon.domain.enumeration;
-
-public enum EventStatus {
-    BEFORE, START, END
-}
