+++ src/main/java/com/personal_project/coupon/coupon/domain/model/CouponIssue.java
@@ -1,6 +1,6 @@
-package com.personal_project.coupon.coupon.domain.entity;
+package com.personal_project.coupon.coupon.domain.model;
 
-import com.personal_project.coupon.coupon.domain.enumeration.CouponIssueStatus;
+import com.personal_project.coupon.coupon.domain.model.enumeration.CouponIssueStatus;
 import com.personal_project.coupon.global.entity.BaseEntity;
 import com.personal_project.coupon.member.domain.Member;
 import jakarta.persistence.*;
