+++ src/main/java/com/personal_project/coupon/coupon/domain/model/enumeration/CouponStatus.java
@@ -1,4 +1,4 @@
-package com.personal_project.coupon.coupon.domain.enumeration;
+package com.personal_project.coupon.coupon.domain.model.enumeration;
 
 public enum CouponStatus {
     PENDING, ISSUED
