+++ src/main/java/com/personal_project/coupon/coupon/domain/model/enumeration/EventStatus.java
@@ -1,0 +1,5 @@
+package com.personal_project.coupon.coupon.domain.model.enumeration;
+
+public enum EventStatus {
+    BEFORE, START, END
+}
