+++ src/main/java/com/personal_project/coupon/coupon/framwork/jpaadapter/CouponAdapter.java
@@ -1,7 +1,7 @@
 package com.personal_project.coupon.coupon.framwork.jpaadapter;
 
 import com.personal_project.coupon.coupon.application.outputport.CouponOutPort;
-import com.personal_project.coupon.coupon.domain.entity.Coupon;
+import com.personal_project.coupon.coupon.domain.model.Coupon;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
