+++ src/main/java/com/personal_project/coupon/coupon/framwork/jpaadapter/CouponIssueAdapter.java
@@ -1,7 +1,7 @@
 package com.personal_project.coupon.coupon.framwork.jpaadapter;
 
 import com.personal_project.coupon.coupon.application.outputport.CouponIssueOutPort;
-import com.personal_project.coupon.coupon.domain.entity.CouponIssue;
+import com.personal_project.coupon.coupon.domain.model.CouponIssue;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
