+++ src/main/java/com/personal_project/coupon/coupon/framwork/jpaadapter/CouponIssueJpaRepository.java
@@ -1,6 +1,6 @@
 package com.personal_project.coupon.coupon.framwork.jpaadapter;
 
-import com.personal_project.coupon.coupon.domain.entity.CouponIssue;
+import com.personal_project.coupon.coupon.domain.model.CouponIssue;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
