+++ src/main/java/com/personal_project/coupon/coupon/framwork/jpaadapter/EventJpaRepository.java
@@ -1,6 +1,6 @@
 package com.personal_project.coupon.coupon.framwork.jpaadapter;
 
-import com.personal_project.coupon.coupon.domain.entity.Event;
+import com.personal_project.coupon.coupon.domain.model.Event;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface EventJpaRepository extends JpaRepository<Event , Long> {
