+++ src/main/java/com/personal_project/coupon/coupon/framwork/web/response/CouponOutPutDTO.java
@@ -1,10 +1,9 @@
 package com.personal_project.coupon.coupon.framwork.web.response;
 
-import com.personal_project.coupon.coupon.domain.entity.Coupon;
-import com.personal_project.coupon.coupon.domain.enumeration.DiscountType;
+import com.personal_project.coupon.coupon.domain.model.Coupon;
+import com.personal_project.coupon.coupon.domain.model.enumeration.DiscountType;
 import lombok.Builder;
 import lombok.Getter;
-import lombok.Setter;
 
 import java.time.LocalDate;
 import java.time.LocalDateTime;
