+++ src/test/java/com/personal_project/coupon/service/CouponIssueTest.java
@@ -3,7 +3,7 @@ package com.personal_project.coupon.service;
 
 import com.personal_project.coupon.coupon.application.service.sync.RedissonCouponIssueService;
 import com.personal_project.coupon.coupon.application.usecase.IssueCouponUsecase;
-import com.personal_project.coupon.coupon.domain.entity.Coupon;
+import com.personal_project.coupon.coupon.domain.model.Coupon;
 import com.personal_project.coupon.coupon.framwork.jpaadapter.CouponAdapter;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,6 +26,7 @@ public class CouponIssueTest {
 
     @Autowired
     private CouponAdapter couponAdapter;
+
 
     @Test
     void 동시에_1000개_요청() throws Exception {
